requirements.txt
Flask==2.3.2
mysql-connector-python==9.4.0
pandas==2.2.3
scikit-learn==1.6.1

